+++ querybook/server/lib/metastore/base_metastore_loader.py
@@ -1,35 +1,39 @@
-from abc import ABCMeta, abstractmethod, abstractclassmethod
-import gevent
 import math
-from typing import List, Dict, Tuple, Optional
 import traceback
+from abc import ABCMeta, abstractclassmethod, abstractmethod
+from typing import Dict, List, Optional, Tuple
 
+import gevent
 from app.db import DBSession, with_session
-from const.metastore import MetadataType, MetastoreLoaderConfig
-from lib.logger import get_logger
-
+from const.metastore import (
+    DataColumn,
+    DataOwnerType,
+    DataTable,
+    MetadataType,
+    MetastoreLoaderConfig,
+)
 from lib.form import AllFormField
+from lib.logger import get_logger
 from lib.utils import json
 from lib.utils.utils import with_exception
-from logic.elasticsearch import update_table_by_id, delete_es_table_by_id
+from logic.elasticsearch import delete_es_table_by_id, update_table_by_id
 from logic.metastore import (
+    create_column,
     create_schema,
-    delete_schema,
     create_table,
-    delete_table,
     create_table_information,
     create_table_ownerships,
-    create_column,
     delete_column,
-    iterate_data_schema,
-    get_table_by_schema_id,
+    delete_schema,
+    delete_table,
     get_column_by_table_id,
     get_schema_by_name,
+    get_table_by_schema_id,
     get_table_by_schema_id_and_name,
+    iterate_data_schema,
 )
-from logic.tag import create_table_tags, create_column_tags
+from logic.tag import create_column_tags, create_table_tags
 
-from .metastore_data_types import DataTable, DataColumn, DataOwnerType
 from .utils import MetastoreTableACLChecker
 
 LOG = get_logger(__name__)
