+++ querybook/server/lib/metastore/loaders/mysql_metastore_loader.py
@@ -1,11 +1,9 @@
 from typing import List, Tuple
 
-from lib.utils.utils import DATETIME_TO_UTC
+from const.metastore import DataColumn, DataTable
 from lib.utils import json as ujson
-from lib.metastore.metastore_data_types import (
-    DataTable,
-    DataColumn,
-)
+from lib.utils.utils import DATETIME_TO_UTC
+
 from .sqlalchemy_metastore_loader import SqlAlchemyMetastoreLoader
